requirements.txt
deepface==0.0.89
numpy==1.23.5
Pillow==10.3.0
streamlit==1.25.0
